requirements.txt
PyGithub >= 1.59
python-dotenv >= 0.20.0
pyyaml >= 6.0.1
jinja2 >= 3.0.3
